Save fixes; work on quest stages
Added an error handler for if subrecord::write_length_prefixed_string writes a string that's too long. If this works (testing is pending), it'll write placeholder text and report an error, though it can't abort the form write on its own.

Fixed QUST/DNAM save code.

Fixed the quest stage UI to properly update the state of controls for flags and forms.

The Papyrus fragment editor can now have maximum lengths set for the script and function names. The quest stage UI makes use of this as appropriate.

+++ native/src/DovahKit/dovah/forms/Quest.cpp
@@ -1371,7 +1371,7 @@ namespace dovah::loaded_forms {
       DNAM.write(this->priority);
       DNAM.skip_bytes(1);
       DNAM.write(this->unknown);
-      DNAM.write(this->quest_type);
+      DNAM.write((uint32_t)this->quest_type);
       DNAM.close();
       //
       if (this->event) {
